video-processing/pyproject.toml
[tool.poetry]
name = "video-processing"
version = "0.1.0"
description = ""
authors = ["skyler arnold <skyarnold1@me.com>"]
readme = "README.md"
packages = [{include = "video_processing"}]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
scipy = "^1.9.3"
numpy = "^1.24.0"
pillow = "^9.3.0"
tensorflow = "^2.11.0"
opencv-python = "^4.6.0.66"
typer = "^0.7.0"
psycopg2 = "^2.9.5"
sqlalchemy = "^1.4.45"
tqdm = "^4.64.1"
# FIXME: depend on mainline pytube once https://github.com/pytube/pytube/pull/1422
# is merged so that the new youtube URL format is supported
pytube = { git = "https://github.com/felipeucelli/pytube", rev = "357a092c63b766db24c2e5060f3c3866dd20a497" }
pandas = "^1.5.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-watch = "^4.2.0"
pytest-mock = "^3.10.0"
ipython = "^8.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
extract-fen = "video_processing.main:app"

[pytest]
log_level = "INFO"

